refactor: :recycle: Add / Remove ModToolStore as a singleton/autoload (#23)
* refactor: :recycle: Add / Remove ModToolStore as a singleton/autoload

Adding the ModToolStore as a singleton on `_enter_tree()` in *plugin.gd* allows easy access from any where in the plugin. Refactored every store reference to use the new singleton. The singleton get's removed in `_exit_tree()` in *plugin.gd*.

* fix: :bug: moved `save_store()` and `load_store()` into *store.gd*

Before `save_store()` was called in `_exit_tree()` in *dock.gd*. Resulting in an error, because the autoload is removed before `_exit_tree()` was executed.

* refactor: :recycle: Added return types