feat: ✨ Mod directory creation from template (#45)
* feat: :truck: moved template files to /templates/default

* feat: :sparkles: Added file dialog to select template dir

* feat: :sparkles: copy template folder

Added `include_empty_dirs := false` to `get_flat_view_dict()` util. This also adds empty dirs to the data. Also added a default value for `path_current_template_dir` in the store.

* feat: :sparkles: Open *mod_main.gd* in script editor

* feat: :sparkles: Update manifest and open manifest editor

* refactor: :recycle: Moved the mod-id split into *create_mod.gd*

the mod name and namespace can be retrieved from the Manifest after it is initialized so no need to store it in the store as well.

* refactor: :recycle: Removed `manifest_dict`

Removed `manifest_dict` completly and manifest_dict_json from the ModToolStore. Updated `update_ui()` to use `.has()` so the `manifest_dict` is no longer needed.