feat: :sparkles: Open Script Editor after script extension creation (#68)
Open the script editor automatically after creating a script extension. This deviates from the planned implementation discussed in #47, as it does not focus on the first selected script but simply opens them in order. This simplification was made to avoid unnecessary complexity. Since in most cases only one script at a time will be extended, the additional complexity is not worth it.

closes #47