feat: :sparkles: updated for Godot 4.1.2
Used the project converter to get a start and fixed all the issues I ran into. Currently, the script files are not updated immediately after using one of the context actions because the `.reload_scripts()` method is missing, and I have not found a workaround for this issue so far.