feat: ✨ updated for Godot 4.1.2 (#104)
* feat: :construction: update for 4.x

* feat: :sparkles: updated for Godot 4.1.2

Used the project converter to get a start and fixed all the issues I ran into. Currently, the script files are not updated immediately after using one of the context actions because the `.reload_scripts()` method is missing, and I have not found a workaround for this issue so far.

* fix: :bug: added back `reload_scripts` as `reload_script`

It's not a full replacement because the editor will show the "newer on disc" popup that can't be hidden via a plugin. To get rid of the popup, the user will have to set the `auto_reload_scripts_on_external_change` editor setting.