refactor: :recycle: reworked `reload_script` again again
Uncommented the switch to the extension script, so `mod_main.gd` is never in focus in the script editor. This means it doesn't need to be instantly refreshed. For overwrites, I added a check to see if `overwrites.gd` is currently open and trigger a refresh if that is the case.